+++ pyobjc-framework-PHASE/PyObjCTest/test_phaseducker.py
@@ -1,0 +1,8 @@
+from PyObjCTools.TestSupport import TestCase
+
+import PHASE
+
+
+class TestPHASEDucker(TestCase):
+    def test_methods(self):
+        self.assertResultIsBOOL(PHASE.PHASEDucker.isActive)