+++ pyobjc-framework-PHASE/PyObjCTest/test_phaseengine.py
@@ -1,0 +1,9 @@
+from PyObjCTools.TestSupport import TestCase
+
+import PHASE
+
+
+class TestPHASEEngine(TestCase):
+    def test_methods(self):
+        self.assertResultIsBOOL(PHASE.PHASEEngine.startAndReturnError_)
+        self.assertArgIsOut(PHASE.PHASEEngine.startAndReturnError_, 0)