+++ pyobjc-framework-PHASE/PyObjCTest/test_phasemixer.py
@@ -1,0 +1,11 @@
+from PyObjCTools.TestSupport import TestCase
+from objc import simd
+
+import PHASE
+
+
+class TestPHASEMixer(TestCase):
+    def test_methods(self):
+        self.assertResultHasType(
+            PHASE.PHASEAmbientMixerDefinition.orientation, simd.simd_quatf.__typestr__
+        )